refactor: exclude the edit string key tests to a separate suite
re #RI-6570
